% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app_kmeans_cluster.R
\name{app_kmeans_cluster}
\alias{app_kmeans_cluster}
\title{app_kmeans_cluster}
\usage{
app_kmeans_cluster(data)
}
\arguments{
\item{data}{data}
}
\description{
A one liner.
}
\details{
Here more details
}
\author{
Reinhard Simon
}

