% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app_view_study.R
\name{app_view_study}
\alias{app_view_study}
\title{app_view_study}
\usage{
app_view_study(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
character a string; default is 'Hello, world!'
}
\description{
Creates a shiny widget to display the study in tabular form.
}
\author{
Reinhard Simon
}

