% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fieldbook.R
\name{get_fieldbook}
\alias{get_fieldbook}
\title{get_fieldbook}
\usage{
get_fieldbook(fieldbook, crop = "", program = "", section = "")
}
\arguments{
\item{crop}{name of crop}

\item{program}{name of breeding program}

\item{section}{section name}

\item{id}{a fieldbook name}
}
\value{
list of fieldbooks
}
\description{
Get a fieldbook.
}
\author{
Reinhard Simon
}
\seealso{
Other res: \code{\link{app_kmeans_cluster}};
  \code{\link{app_view_study}}
}

