% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_study.R
\name{get_study}
\alias{get_study}
\title{get_study}
\usage{
get_study(study, section = c("Fieldbook"))
}
\arguments{
\item{study}{path to study}

\item{section}{section name}
}
\value{
data.frame
}
\description{
Get a section (=sheet) from a study.
}
\author{
Reinhard Simon
}
\seealso{
Other study: \code{\link{list_studies}}
}

