% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_studies.R
\name{list_studies}
\alias{list_studies}
\title{list_studies}
\usage{
list_studies(crop = "potato", database = "D:/DataCollector/DC4RT/data")
}
\arguments{
\item{crop}{a crop name}

\item{database}{path}
}
\value{
list of study names and access paths
}
\description{
List all studies for a given crop and database.
}
\details{
Currently returns a list of short hand name for display and full path name. The latter
may change in a future revision.
}
\author{
author
}
\seealso{
Other study: \code{\link{get_study}}
}

